+++ test/pch.hpp
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: (c) 2026 Shawn Silverman <shawn@pobox.com>
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+// pch.hpp exists so that Boost compiles.
+// This file is part of the QNEthernet library.
+
+#pragma once